some more cleanup and splitting of code
