+++ tools/server-spring/src/main/java/org/eclipse/rdf4j/http/server/repository/AbstractRepositoryController.java
@@ -1,10 +1,10 @@
 /*******************************************************************************
- * Copyright (c) 2015 Eclipse RDF4J contributors, Aduna, and others.
+ * Copyright (c) 2022 Eclipse RDF4J contributors.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Distribution License v1.0
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/org/documents/edl-v10.php.
- *******************************************************************************/
+ ******************************************************************************/
 package org.eclipse.rdf4j.http.server.repository;
 
 import javax.servlet.http.HttpServletRequest;
@@ -30,16 +30,22 @@ public abstract class AbstractRepositoryController extends AbstractController {
 	protected ModelAndView handleRequestInternal(HttpServletRequest request, HttpServletResponse response)
 			throws Exception {
 
+		logger.debug("Request method: {}", request.getMethod());
+
 		RequestMethod requestMethod = RequestMethod.valueOf(request.getMethod());
 
 		switch (requestMethod) {
 		case DELETE: {
+			logger.debug("handleDeleteRepositoryRequest");
 			return getRepositoryRequestHandler().handleDeleteRepositoryRequest(request);
 		}
 		case PUT: {
+			logger.debug("handleCreateNewRepositoryRequest");
 			return getRepositoryRequestHandler().handleCreateNewRepositoryRequest(request);
 		}
 		}
+
+		logger.debug("handleQueryRequest");
 
 		return getQueryRequestHandler().handleQueryRequest(request, requestMethod, response);
 	}
